+++ Pandora/Paradigm/Structure/Ability/Substructure.hs
@@ -7,7 +7,7 @@ import Pandora.Pattern.Category ((.))
 import Pandora.Pattern.Functor.Covariant (comap)
 import Pandora.Pattern.Functor.Extractable (extract)
 import Pandora.Paradigm.Controlflow.Effect.Interpreted (run)
-import Pandora.Paradigm.Inventory.Optics (type (:~.), view, set)
+import Pandora.Paradigm.Inventory.Optics (type (:~.), view)
 import Pandora.Paradigm.Primary.Functor.Tagged (Tagged (Tag))
 import Pandora.Paradigm.Schemes.TU (TU (TU), type (<:.>))
 
@@ -20,8 +20,5 @@ class Substructure f t where
 
 	subview :: t ~> Substructural f t
 	subview = view (sub @f)
-
-	subplace :: (Substructural f t) a -> t a -> t a
-	subplace = set (sub @f)
 
 data Segment a = Tail a